pll.c: 1: yes: MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012
Fixed MISRA 10.7 violation where composite expression had operands of
different essential type widths by explicitly casting uint32_t division
result to uint64_t before addition to fret variable. 1 violation fixed,
ID: MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012, ensures type compatibility
in arithmetic operations.

+++ plat/ti/k3low/common/drivers/pm/clock/pll.c
@@ -233,7 +233,7 @@ static enum consider_result pll_consider(struct pll_consider_data *data,
 			/* Calculate fractional part of frequency */
 			fret = (uint64_t)(data->input / rem_div) * curr_pllfm;
 			frem = (uint64_t)(data->input % rem_div) * curr_pllfm;
-			fret += (uint32_t)frem / rem_div;
+			fret += (uint64_t)((uint32_t)frem / rem_div);
 			frem = (uint32_t)frem % rem_div;
 
 			/* Fold in multiplier */
